+++ packages/grid/src/features/standard-table/components/StandardTableRowList.tsx
@@ -1,7 +1,6 @@
 import * as React from "react";
 import { useMemo } from "react";
 import { InfiniteList } from "../../table-ui/components/InfiniteList";
-import { tableRowHeightPixels } from "../../../config/TableConfig";
 import { useColumnValueResolver } from "../hooks/UseColumnValueResolver";
 import {
   useStandardTableConfig,
@@ -9,17 +8,21 @@ import {
 } from "../hooks/UseStandardTableConfig";
 import { multitypeComparator } from "../util/MultitypeComparator";
 import { StandardTableRow } from "./StandardTableRow";
+import { elementHeightPerVariant } from "../config/StandardTableInfiniteConfig";
+import { StandardTableVariant } from "./StandardTable";
 
 interface StandardTableContentProps<TItem> {
   items?: Array<TItem>;
   colIndexOffset?: number;
   rowIndexOffset?: number;
+  variant: StandardTableVariant;
 }
 
 export const StandardTableRowList = React.memo(function StandardTableRowList<
   TItem
 >({
   items,
+  variant,
   colIndexOffset = 0,
   rowIndexOffset = 0,
 }: StandardTableContentProps<TItem>) {
@@ -56,7 +59,7 @@ export const StandardTableRowList = React.memo(function StandardTableRowList<
     <InfiniteList
       disabled={disableInfiniteList || enableExpandCollapse}
       length={sortedItems.length}
-      elementHeight={tableRowHeightPixels}
+      elementHeight={elementHeightPerVariant[variant]}
       threshold={30}
     >
       {sortedItems.map((item, index) => (