+++ packages/grid/src/features/standard-table/components/TextCell.tsx
@@ -1,5 +1,6 @@
 import { Indent, Text } from "@stenajs-webui/core";
 import * as React from "react";
+import styles from "./TextCell.module.css";
 
 interface Props {
   label?: string;
@@ -7,8 +8,10 @@ interface Props {
 
 export const TextCell: React.FC<Props> = React.memo(function ({ label }) {
   return (
-    <Indent>
-      <Text>{label}</Text>
+    <Indent overflow={"hidden"}>
+      <Text className={styles.textCell} title={label}>
+        {label}
+      </Text>
     </Indent>
   );
 });